// data-integration-ui/src/features/projects/projects.types.ts
export type Properties = { [key: string]: ItemPropertiesResponse }
export interface ItemResponse {
    id: string
    properties: Properties
    mappingIds: string[]
}

export interface ItemPropertiesResponse {
    value: string
    originalValue?: string
}
// ...
export interface GetScopeHeadersResponse {
    headers: string[]
    extraHeaders: string[]
}